requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
dj-database-url==2.3.0
dj3-cloudinary-storage==0.0.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.9.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
